# openapi.yaml
openapi: 3.0.3
info:
  title: Alt-text Backend API
  description: |-
    This is the Alt-text Backend API based on the OpenAPI 3.0 specification.
  contact:
    email: da.cruz@aol.com
  version: 1.0.11
...
paths:
  /books:
    get:
      tags:
        - Books
      summary: Get a list of books.
      description: Get a list of book objects given various options.
      operationId: getBooks
      parameters:
        - name: titleQ
          in: query
          description: String to match the title to.
          required: false
          explode: true
          schema:
            type: string
        - name: limit
          in: query
          description: Max number of books to return.
          required: false
          explode: true
          schema:
            type: integer
        - name: skip
          in: query
          description: Number of books to skip.
          required: false
          explode: true
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
        '500':
          description: Internal Server Error
    post:
      tags:
        - Books
      summary: Add a book.
      description: Submit a new book for analyzation and store results in database.
      operationId: addBook
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                id:
                  type: string
                  description: Id of the book (optional).
                title:
                  type: string
                  description: Title of the book.
                book:
                  type: string
                  description: Zip file of the book.
                  format: binary
                cover:
                  type: string
                  description: Cover image for the book (optional).
                  format: binary
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '500':
          description: Internal Server Error
  /books/{bookid}:
    parameters:
      - name: bookid
        in: path
        description: Id of the book.
        required: true
        explode: true
        schema:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
    get:
      tags:
        - Books
      summary: Get a book.
      description: Get a book object with basic information by its id.
      operationId: getBook
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '500':
          description: Internal Server Error
    patch:
      tags:
        - Books
      summary: Update a book's metadata.
      description: Update a book's metadata by its id.
      operationId: updateBook
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Title of the book (optional).
                cover:
                  type: string
                  description: Cover image for the book (optional).
                  format: binary
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
            application/xml:
              schema:
                $ref: '#/components/schemas/Book'
        '500':
          description: Internal Server Error
    put:
      tags:
      - Books
      summary: Re-analyze an entire book.
      description: Re-analyze an entire book and overwrite current image data by its id.
      operationId: analyzeBook
      parameters:
        - name: missingOnly
          in: query
          description: If analyzing on upload, whether to analyze only the images without alt-text.
          required: false
          explode: true
          schema:
            type: boolean
            example: true
            default: true
        - name: waitForAnalysis
          in: query
          description: Whether to wait for the analysis to complete before returning a response.
          required: false
          explode: true
          schema:
            type: boolean
            example: false
            default: false
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '500':
          description: Internal Server Error
    delete:
      tags:
        - Books
      summary: Delete a book.
      description: Delete a book by its id.
      operationId: deleteBook
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '500':
          description: Internal Server Error
...
components:
  schemas:
    Book:
      type: object
      properties:
        id:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        title:
          type: string
          example: "Diary of an Oxygen Thief"
        size:
          type: string
          example: "1.16MB"
        status:
          type: string
          example: "processing"
          enum: ["available", "processing", "deleted"]
          default: "available"
        numImages:
          type: integer
          example: 4
...